extract `<APIDocs />` in `components/api-docs.tsx`
